removed unwanted classes and parameters
