+++ core/api/src/androidMain/kotlin/com/gu/recipe/api/di/FeastAndroidApiModule.kt
@@ -3,8 +3,8 @@ package com.gu.recipe.api.di
 import android.content.Context
 import com.gu.recipe.api.config.FeastApiConfig
 import com.gu.recipe.api.config.toGraphQlConfig
-import com.gu.recipe.api.repository.GraphQlRecipeRepository
-import com.gu.recipe.api.repository.RecipeRepository
+import com.gu.recipe.api.repository.GraphQlRepositoryImpl
+import com.gu.recipe.api.repository.GraphQLRepository
 import com.gu.recipe.core.graphql.androidGraphQlModule
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
@@ -16,7 +16,7 @@ fun androidFeastApiModule(
     ioDispatcher: CoroutineDispatcher = Dispatchers.IO,
 ): Module = module {
     includes(androidGraphQlModule(config.toGraphQlConfig(), ioDispatcher))
-    single<RecipeRepository> { GraphQlRecipeRepository(get()) }
+    single<GraphQLRepository> { GraphQlRepositoryImpl(get()) }
 }
 
 fun androidFeastApiModule(
@@ -24,6 +24,6 @@ fun androidFeastApiModule(
     config: FeastApiConfig,
 ): Module = module {
     includes(androidGraphQlModule(context, config.toGraphQlConfig()))
-    single<RecipeRepository> { GraphQlRecipeRepository(get()) }
+    single<GraphQLRepository> { GraphQlRepositoryImpl(get()) }
 }
