+++ core/api/src/commonMain/kotlin/com/gu/recipe/api/di/FeastApiModule.kt
@@ -2,8 +2,8 @@ package com.gu.recipe.api.di
 
 import com.gu.recipe.api.config.FeastApiConfig
 import com.gu.recipe.api.config.toGraphQlConfig
-import com.gu.recipe.api.repository.GraphQlRecipeRepository
-import com.gu.recipe.api.repository.RecipeRepository
+import com.gu.recipe.api.repository.GraphQlRepositoryImpl
+import com.gu.recipe.api.repository.GraphQLRepository
 import com.gu.recipe.core.graphql.di.graphQlModule
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
@@ -15,6 +15,6 @@ fun feastApiModule(
     ioDispatcher: CoroutineDispatcher = Dispatchers.Default,
 ): Module = module {
     includes(graphQlModule(config.toGraphQlConfig(), ioDispatcher))
-    single<RecipeRepository> { GraphQlRecipeRepository(get()) }
+    single<GraphQLRepository> { GraphQlRepositoryImpl(get()) }
 }
