--- core/api/src/commonMain/kotlin/com/gu/recipe/api/models/FrontResponse.kt
@@ -1,8 +1,0 @@
-package com.gu.recipe.api.models
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class FrontResponse(
-    val id: String,
-)
