+++ core/api/src/commonMain/kotlin/com/gu/recipe/api/repository/GraphQLRepository.kt
@@ -1,13 +1,14 @@
 package com.gu.recipe.api.repository
 
+import com.gu.recipe.core.graphql.GraphQlResult
 import com.gu.recipe.core.graphql.generated.GetFrontsByRegionQuery
 import com.gu.recipe.core.graphql.generated.type.Editions
 import com.gu.recipe.core.graphql.generated.type.Regions
 
-interface RecipeRepository {
+interface GraphQLRepository {
     suspend fun getFrontByRegion(
         region: Regions,
         edition: Editions,
         recipesLimit: Int
-    ): Result<List<GetFrontsByRegionQuery.Front>>
+    ): GraphQlResult<List<GetFrontsByRegionQuery.Front>>
 }
