--- core/api/src/commonMain/kotlin/com/gu/recipe/api/repository/GraphQlRecipeRepository.kt
@@ -1,47 +1,0 @@
-package com.gu.recipe.api.repository
-
-import com.gu.recipe.core.graphql.GraphQlError
-import com.gu.recipe.core.graphql.GraphQlResult
-import com.gu.recipe.core.graphql.generated.GetFrontsByRegionQuery
-import com.gu.recipe.core.graphql.generated.type.Editions
-import com.gu.recipe.core.graphql.generated.type.Regions
-import com.gu.recipe.core.graphql.repository.RecipeGraphQlDataSource
-
-class GraphQlRecipeRepository(
-    private val dataSource: RecipeGraphQlDataSource,
-) : RecipeRepository {
-
-    override suspend fun getFrontByRegion(
-        region: Regions,
-        edition: Editions,
-        recipesLimit: Int
-    ): Result<List<GetFrontsByRegionQuery.Front>> {
-        return when (val result = dataSource.getFrontByRegion(
-            region = region,
-            edition = edition,
-            recipesLimit = recipesLimit,
-        )) {
-            is GraphQlResult.Success -> {
-                val front = result.value
-                Result.success(front)
-            }
-
-            is GraphQlResult.Failure -> Result.failure(result.error.toRepositoryError())
-        }
-    }
-
-    private fun GraphQlError.toRepositoryError(): Exception = when (this) {
-        is GraphQlError.GraphQl -> RecipeRepositoryError.GraphQl(messages)
-        is GraphQlError.Transport -> RecipeRepositoryError.Network(cause)
-        is GraphQlError.Unexpected -> RecipeRepositoryError.Unexpected(cause)
-        GraphQlError.MissingData -> RecipeRepositoryError.MissingData
-    }
-}
-
-sealed class RecipeRepositoryError : Exception() {
-    data class GraphQl(val messages: List<String>) : Exception()
-    data class Network(override val cause: Throwable) : Exception()
-    data class Unexpected(override val cause: Throwable) : Exception()
-    object MissingData : Exception()
-}
-
