+++ core/api/src/iosMain/kotlin/com/gu/recipe/api/FeastIos.kt
@@ -2,7 +2,7 @@ package com.gu.recipe.api
 
 import com.gu.recipe.api.config.FeastApiConfig
 import com.gu.recipe.api.di.iosFeastApiModule
-import com.gu.recipe.api.repository.RecipeRepository
+import com.gu.recipe.api.repository.GraphQLRepository
 import com.gu.recipe.core.networking.FeastNetworkApiEndpoint
 import com.gu.recipe.core.networking.NetworkConfig
 import org.koin.core.context.startKoin
@@ -12,12 +12,10 @@ import org.koin.mp.KoinPlatform
 object FeastIos {
     fun start(
         endpoint: FeastNetworkApiEndpoint,
-        serverUrl: String? = null,
     ) {
         start(
             FeastApiConfig(
                 networkConfig = NetworkConfig(endpoint = endpoint),
-                serverUrl = serverUrl,
             ),
         )
     }
@@ -29,7 +27,7 @@ object FeastIos {
         }
     }
 
-    fun recipeRepository(): RecipeRepository = KoinPlatform.getKoin().get()
+    fun recipeRepository(): GraphQLRepository = KoinPlatform.getKoin().get()
 
     fun stop() {
         KoinPlatform.getKoinOrNull()?.let { stopKoin() }
