+++ core/api/src/iosMain/kotlin/com/gu/recipe/api/di/FeastIosApiModule.kt
@@ -2,8 +2,8 @@ package com.gu.recipe.api.di
 
 import com.gu.recipe.api.config.FeastApiConfig
 import com.gu.recipe.api.config.toGraphQlConfig
-import com.gu.recipe.api.repository.GraphQlRecipeRepository
-import com.gu.recipe.api.repository.RecipeRepository
+import com.gu.recipe.api.repository.GraphQlRepositoryImpl
+import com.gu.recipe.api.repository.GraphQLRepository
 import com.gu.recipe.core.graphql.iosGraphQlModule
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
@@ -15,6 +15,6 @@ fun iosFeastApiModule(
     ioDispatcher: CoroutineDispatcher = Dispatchers.Default,
 ): Module = module {
     includes(iosGraphQlModule(config.toGraphQlConfig(), ioDispatcher))
-    single<RecipeRepository> { GraphQlRecipeRepository(get()) }
+    single<GraphQLRepository> { GraphQlRepositoryImpl(get()) }
 }
