+++ core/graphql/src/commonMain/kotlin/com/gu/recipe/core/graphql/repository/RecipeGraphQlDataSource.kt
@@ -5,7 +5,6 @@ import com.gu.recipe.core.graphql.generated.GetFrontsByRegionQuery
 import com.gu.recipe.core.graphql.generated.type.Editions
 import com.gu.recipe.core.graphql.generated.type.Regions
 
-// TODO: remove this interface, not required
 interface RecipeGraphQlDataSource {
     suspend fun getFrontByRegion(
         region: Regions,
